fixed gc overhead issue in downaloder and improved workflow
Former-commit-id: 06c0322165c853d5a44d8eaec34fc512806c12b8

+++ backend/services/downloader/src/main/scala/edu/ucdavis/fiehnlab/mona/backend/services/downloader/service/DownloaderService.scala
@@ -86,7 +86,7 @@ class DownloaderService extends LazyLogging {
     val exportFile: Path = Paths.get(exportDir, exportFilename)
     val compressedFile: Path = Paths.get(exportDir, compressedExportFilename)
 
-    val count: Int = downloadWriterService.writeExportFile(exportFile, compressedFile, export.query, export.format, compressExport)
+    val count: Long = downloadWriterService.writeExportFile(exportFile, compressedFile, export.query, export.format, compressExport)
 
     if (compressExport) {
       // Update export object with filesize, query count and filenames
